카드 추가, 삭제 구현

+++ BE/src/main/java/com/codesquad/todo/domain/ActivityRepository.java
@@ -2,5 +2,5 @@ package com.codesquad.todo.domain;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface ActivityRepository extends CrudRepository<Activity, Integer> {
+public interface ActivityRepository extends CrudRepository<Activity, Long> {
 }
